fumount: Add a way to forcibly umount a filesystem
This feature lets you unmount a filesytem that is in use.  Processes
holding file descriptors will get EBADF on any access to the file.
Processes holding maps will get a SIGSEGV if they try to access that
memory.  Processes with their root directory or cwd in the filesystem
will be booted to /.

Signed-off-by: Corey Minyard <cminyard@mvista.com>